requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
paho-mqtt
pymongo
pillow
opencv-python
vietocr
roboflow
numpy